move ignored tests to name_resolver integration test
