+++ crates/name_resolver/Cargo.toml
@@ -21,3 +21,7 @@ gem_evm = { path = "../gem_evm" }
 
 [dev-dependencies]
 tokio-test = { version = "*" }
+
+[[test]]
+name = "integration_test"
+test = false
