+++ crates/name_resolver/README.md
@@ -1,0 +1,29 @@
+# Name Resolver
+
+## Supported Name Providers
+
+- [x] [Ethereum Name Service](https://ens.domains/)
+- - .eth
+- [x] [Base Name Service](https://www.basename.app/)
+- - .base
+- [x] [Unstoppable Domains](https://unstoppabledomains.com/)
+- - .crypto
+- [x] [Solana Name Service](https://www.sns.id/)
+- - .so
+- [x] [TON Domains](https://dns.ton.org/)
+- - .ton
+- [x] [Aptos Names](https://www.aptosnames.com/)
+- - .apt
+- [x] [d.id](https://d.id/)
+- - .bit
+- [x] [Interchain Nameservice](https://www.icns.xyz/)
+- - .atom
+- - .osmo
+- [x] [Lens](https://www.lens.xyz/)
+- - .lens
+- [x] [eths.center](https://eths.center/)
+- - .tree
+- [x] [Space ID](https://space.id)
+- - .bnb
+- [x] [Sui Name Service](https://suins.io/)
+- - .sui
