+++ crates/name_resolver/src/base.rs
@@ -51,23 +51,3 @@ impl NameClient for BNSClient {
         vec![Chain::Base]
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tokio_test::block_on;
-
-    #[ignore]
-    #[test]
-    fn test_resolve() {
-        // this test is ignored from UT casue it connects to the real network
-        block_on(async {
-            let client = super::BNSClient::new(String::from("https://resolver-api.basename.app"));
-            let addres = client.resolve("hello.base", primitives::Chain::Base).await;
-            assert_eq!(
-                addres.unwrap(),
-                "0x4fb3f133951bF1B2d52fF6CEab2c703fbB6E98cC"
-            )
-        });
-    }
-}
