+++ crates/name_resolver/src/ens_provider/provider.rs
@@ -36,24 +36,3 @@ impl Provider {
         todo!()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::Provider;
-    use primitives::Chain;
-    use tokio_test::block_on;
-
-    #[ignore]
-    #[test]
-    fn test_resolver() {
-        // this test is ignored from UT casue it connects to the real network
-        block_on(async {
-            let provider = Provider::new(String::from("https://eth.llamarpc.com"));
-            let addres = provider.resolve_name("vitalik.eth", Chain::Ethereum).await;
-            assert_eq!(
-                addres.unwrap(),
-                "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045".to_lowercase()
-            )
-        });
-    }
-}
